[FE Style] 이미지 컴포넌트 분리및 스타일링
